+++ Runtime/Math/Matrix3.cs
@@ -1,9 +1,11 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) 2017-2021 Juan Delgado (@JuDelCo)
 
+using System;
+
 namespace Ju.Math
 {
-	public partial struct Matrix3
+	public partial struct Matrix3 : IEquatable<Matrix3>
 	{
 		public float m00;
 		public float m01;
@@ -152,6 +154,11 @@ namespace Ju.Math
 				hash = (hash * 23) ^ m22.GetHashCode();
 				return hash;
 			}
+		}
+
+		public bool Equals(Matrix3 other)
+		{
+			return (this == other);
 		}
 
 		public override bool Equals(object obj)