+++ Runtime/Math/Quat.cs
@@ -1,9 +1,11 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) 2017-2021 Juan Delgado (@JuDelCo)
 
+using System;
+
 namespace Ju.Math
 {
-	public partial struct Quat
+	public partial struct Quat : IEquatable<Quat>
 	{
 		public float x;
 		public float y;
@@ -112,6 +114,11 @@ namespace Ju.Math
 				hash = hash * 23 + w.GetHashCode();
 				return hash;
 			}
+		}
+
+		public bool Equals(Quat other)
+		{
+			return (this == other);
 		}
 
 		public override bool Equals(object obj)