+++ Runtime/Math/Vector2f.cs
@@ -1,9 +1,11 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) 2017-2021 Juan Delgado (@JuDelCo)
 
+using System;
+
 namespace Ju.Math
 {
-	public partial struct Vector2f
+	public partial struct Vector2f : IEquatable<Vector2f>
 	{
 		public float x;
 		public float y;
@@ -168,6 +170,11 @@ namespace Ju.Math
 				hash = hash * 23 + y.GetHashCode();
 				return hash;
 			}
+		}
+
+		public bool Equals(Vector2f other)
+		{
+			return (this == other);
 		}
 
 		public override bool Equals(object obj)