+++ Runtime/Physics/FloatRect.cs
@@ -1,9 +1,11 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) 2017-2021 Juan Delgado (@JuDelCo)
 
+using System;
+
 namespace Ju.Math
 {
-	public partial struct FloatRect
+	public partial struct FloatRect : IEquatable<FloatRect>
 	{
 		public float x;
 		public float y;
@@ -116,6 +118,11 @@ namespace Ju.Math
 				hash = hash * 23 + height.GetHashCode();
 				return hash;
 			}
+		}
+
+		public bool Equals(FloatRect other)
+		{
+			return (this == other);
 		}
 
 		public override bool Equals(object obj)