+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 
 # Changelog
 
+## [1.8.0] - 2021-05-31
+
+### Added
+
+- Add two new LerpRounded methods.
+
+### Improved
+
+- Add IEquatable interface to all structs.
+
 ## [1.7.0] - 2021-05-24
 
 ### Added
@@ -13,4 +23,5 @@
 - Update .editorconfig to include generated folders by dotnet compiler.
 - Update .gitignore to exclude generated folders by dotnet compiler.
 
+[1.8.0]: https://github.com/JuDelCo/CoreMath/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/JuDelCo/CoreMath/compare/v1.6.0...v1.7.0